+++ freesynd/trunk/src/vehicle.cpp
@@ -509,7 +509,7 @@ bool VehicleInstance::movementV(int elapsed)
         } else {
             setDirection(diffx, diffy, &dir_);
             int dx = 0, dy = 0;
-            float d = sqrt((float) (diffx * diffx + diffy * diffy));
+            double d = sqrt((double)(diffx * diffx + diffy * diffy));
 
             if (abs(diffx) > 0)
                 // dx = diffx * (speed_ * elapsed / 1000) / d;
